+++ sdk/messaging/azservicebus/internal/atom/entity_manager.go
@@ -172,7 +172,7 @@ func (em *entityManager) execute(ctx context.Context, method string, entityPath 
 	}
 
 	q := req.Raw().URL.Query()
-	q.Add("api-version", "2021-05")
+	q.Add("api-version", "2024-05")
 	req.Raw().URL.RawQuery = q.Encode()
 
 	if body != nil {